fix(audit): pass 1 — settle-defer hardening and routing TU relocation
- completeDeferredWindowRoutes: re-derive the structural eligibility pair
  against the settled flags before the snap arms consume canSnapRestore;
  the defer-time verdict predates a same-burst keep-above/skip-switcher.
- reevaluateWindowEligibility: carry the scrolling-fullscreen exemption
  isEligibleForTilingNotify uses, so a benign flag edge on a fullscreen
  scrolling tile does not evict it; release first-frame suppression after
  an eviction (the spawn restore re-lands on spawnGeometry, so the
  geometry-settle release could never fire).
- Move deferWindowRouting / reevaluateWindowEligibility /
  completeDeferredWindowRoutes to floatcleanup.cpp (the eligibility TU)
  so the grandfathered tilinghandler.cpp overrun shrinks instead of
  growing (1408 -> 1292).
- Drop the now-callerless isScreenQueryPending() accessor; refresh the
  stale tryInstantSnapRestore caller comment.
